Add PersonalizedTreatmentSummary data structure and related classes
