--- datamodel/src/main/kotlin/com/hartwig/actin/datamodel/algo/SimilarPatientsSummary.kt
@@ -1,7 +1,0 @@
-package com.hartwig.actin.datamodel.algo
-
-data class SimilarPatientsSummary(
-    val overallTreatmentProportion: List<TreatmentProportion>,
-    val similarPatientsTreatmentProportion: List<TreatmentProportion>
-) {
-}
