--- datamodel/src/main/kotlin/com/hartwig/actin/datamodel/algo/TreatmentEfficacyPrediction.kt
@@ -1,8 +1,0 @@
-package com.hartwig.actin.datamodel.algo
-
-data class TreatmentEfficacyPrediction(
-    val treatment: String,
-    val survivalProbs: List<Double>,
-    val shapValues: Map<String, ShapDetail>
-) {
-}
