--- datamodel/src/main/kotlin/com/hartwig/actin/datamodel/algo/TreatmentProportion.kt
@@ -1,6 +1,0 @@
-package com.hartwig.actin.datamodel.algo
-
-data class TreatmentProportion(
-    val treatment: String,
-    val proportion: Double
-)
